# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brewed Bliss Café API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: string;
  category: string;
  image: string;
}

export interface CafeImage {
  id: string;
  category: string;
  objectPath: string;
  label: string;
  uploadedAt: string;
}

export interface RegisterImageInput {
  category: string;
  objectPath: string;
  label: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type GetImagesParams = {
category?: string;
};