# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Brewed Bliss Café API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form operations
  - name: menu
    description: Menu operations
  - name: images
    description: Image management operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Submits a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /menu:
    get:
      operationId: getMenu
      tags: [menu]
      summary: Get all menu items
      description: Returns all menu items grouped by category
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /images:
    get:
      operationId: getImages
      tags: [images]
      summary: Get images by category
      description: Returns images filtered by category, or all images if no category provided
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CafeImage"

  /images/register:
    post:
      operationId: registerImage
      tags: [images]
      summary: Register an uploaded image
      description: Saves image metadata after a successful object storage upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterImageInput"
      responses:
        "200":
          description: Image registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CafeImage"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /images/{id}:
    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete an image
      description: Removes image metadata (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset from PUBLIC_OBJECT_SEARCH_PATHS
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: File not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve public object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ContactInput:
      type: object
      required:
        - name
        - email
        - message
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        message:
          type: string
          minLength: 1

    ContactResult:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string

    MenuItem:
      type: object
      required:
        - id
        - name
        - description
        - price
        - category
        - image
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: string
        category:
          type: string
        image:
          type: string

    CafeImage:
      type: object
      required:
        - id
        - category
        - objectPath
        - label
        - uploadedAt
      properties:
        id:
          type: string
        category:
          type: string
        objectPath:
          type: string
        label:
          type: string
        uploadedAt:
          type: string

    RegisterImageInput:
      type: object
      required:
        - category
        - objectPath
        - label
      properties:
        category:
          type: string
        objectPath:
          type: string
        label:
          type: string

    DeleteResult:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean

    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"