# lib/api-zod/src/generated/types/registerImageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brewed Bliss Café API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterImageInput {
  category: string;
  objectPath: string;
  label: string;
}